requirements.txt
openai>=1.54.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pymongo>=4.6.0
